// src/types.ts
export type Page = {
  TITLE: string
  DESCRIPTION: string
}

export type Site = {
  TITLE: string
  DESCRIPTION: string
}

export type Links = {
  TEXT: string
  HREF: string
}[]

export type Info = {
  PHONE: string
  PHONE_HREF: string
  EMAIL: string
  EMAIL_HREF: string
  // ADDRESS: string
}

export type Socials = {
  NAME: string
  ICON: string
  TEXT: string
  HREF: string
}[]